Add files via upload
Lots of work, added a ton of sections, fixed a couple weird bugs like when I would click a button to smooth scroll it would focus on the section and add a thin black border around it, so thats gone now. Also the fitness classes part was tricky, instead of using clip-path polygon, which what I was trying to do at first, I just photoshopped the photo to have the slice effect. That way I would still have the border to give it a clean look.